refactor: extract types and rename variables
