+++ src/views/Settings/Settings.tsx
@@ -2,7 +2,8 @@ import { View } from '@/components/Layout/View'
 import { Button, ScrollArea } from '@radix-ui/themes'
 import { ProxySettings } from './ProxySettings'
 import { FormProvider, useForm } from 'react-hook-form'
-import { AppSettings, AppSettingsSchema } from '@/schemas/appSettings'
+import { AppSettingsSchema } from '@/schemas/appSettings'
+import { AppSettings } from '@/types/settings'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
