+++ src/views/Settings/UpstreamProxySettings.tsx
@@ -1,5 +1,5 @@
 import { FieldGroup } from '@/components/Form'
-import { AppSettings } from '@/schemas/appSettings'
+import { AppSettings } from '@/types/settings'
 import { TextField, Flex, Checkbox, Text } from '@radix-ui/themes'
 import { Controller, useFormContext } from 'react-hook-form'
